Self-implemented QUBO and QAOA for the SatelliteSolver (#152)
This PR marks the first step of updating the MQT ProblemSolver to
support Qiskit >= 2.0.0. For this, the dependencies
'qiskit_optimization' and 'qiskit_algorithms' have to be removed. To
this end, the QUBO to QAOA process in the SatelliteSolver has been
implemented manually. Additionally, small fixes have been made to
support Qiskit >= 2.0.0.

---------

Signed-off-by: Tobias Forster <158472669+tobi-forster@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>